backend/requirements.txt
firebase-admin
google-auth
google-auth-oauthlib
google-auth-httplib2
fastapi
uvicorn[standard]
pydantic
python-multipart
sentence-transformers
transformers
faiss-cpu
numpy
torch  
pdfplumber
camelot-py[cv]
python-docx
PyMuPDF
pytesseract
Pillow
nltk
regex
redis
aioredis
python-dotenv
pydantic-settings
pytest
pytest-asyncio
httpx


